Remove word boundary from phone_regex2
Noticed that (555)555-1234 does not match completely in the regex as-is, but returns `555)555-1234`.

This is a trivial formatting difference, and likely of no impact to workflows.

The cause for this is that `\b` matches the transition from a `\w` word character to a `[^\w]` non-word character. Numbers are word characters, but `(` is not, so the required `\b` matches the position between the `([matches here]5`, and the `\(?` cannot be matched there.